sageCoach/requirements.txt
Flask
flask-cors
Werkzeug
numpy
opencv-python
Pillow
mediapipe
google-generativeai
openai
firebase-admin
